# services/booking-service/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para reservas, asientos, checkout y tickets.",
        "title": "SeatGuard Booking Service API",
        "contact": {
            "name": "Lucas Cabral",
            "email": "lucassimple@hotmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/api/v1",
    "paths": {
...
        "/checkout": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene todos los registros de checkout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checkout"
                ],
                "summary": "Obtener todos los checkouts",
                "responses": {
                    "200": {
                        "description": "Checkouts encontrados",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Checkout"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualiza un registro de checkout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checkout"
                ],
                "summary": "Actualizar checkout",
                "parameters": [
                    {
                        "description": "Datos del checkout",
                        "name": "checkout",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Checkout"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checkout actualizado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/models.Checkout"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea un nuevo registro de checkout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checkout"
                ],
                "summary": "Crear checkout",
                "parameters": [
                    {
                        "description": "Datos del checkout",
                        "name": "checkout",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Checkout"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Checkout creado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/models.Checkout"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/checkout/{orderID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene un registro de checkout por order ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checkout"
                ],
                "summary": "Obtener checkout por order ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "orderID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checkout encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.Checkout"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.BookingOrder": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "eventHour": {
                    "type": "string"
                },
                "eventName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Seat"
                    }
                },
                "paymentProviderId": {
                    "description": "Token o ID de transacción de la pasarela de pago (Stripe/MercadoPago)",
                    "type": "string"
                },
                "seatIds": {
                    "description": "Asientos involucrados en esta orden\nSeatIDs []string `gorm:\"type:text[]\" json:\"seatIds\"`",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "$ref": "#/definitions/models.PaymentStatus"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "models.Checkout": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "customerEmail": {
                    "type": "string"
                },
                "customerId": {
                    "type": "string"
                },
                "customerName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "order": {
                    "$ref": "#/definitions/models.BookingOrder"
                },
                "orderId": {
                    "type": "string"
                },
                "paymentIntentId": {
                    "type": "string"
                },
                "paymentProvider": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.PaymentStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "COMPLETED",
                "FAILED"
            ],
            "x-enum-varnames": [
                "PaymentPending",
                "PaymentCompleted",
                "PaymentFailed"
            ]
        },
        "models.Seat": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "eventHour": {
                    "type": "string"
                },
                "eventId": {
                    "description": "Relaciones",
                    "type": "string"
                },
                "eventName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lockedAt": {
                    "description": "Cuando se desbloquea",
                    "type": "string"
                },
                "lockedBy": {
                    "description": "Control de Bloqueo Temporal",
                    "type": "string"
                },
                "number": {
                    "description": "\"10\", \"A1\", etc.",
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "section": {
                    "description": "Identificación del asiento (ej: A-10, B-5)",
                    "type": "string"
                },
                "status": {
                    "description": "Estado actual del asiento\nenums: AVAILABLE, RESERVED, SOLD, BLOCKED",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.SeatStatus"
                        }
                    ]
                },
                "ticketId": {
                    "description": "ID del ticket final si se vende",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.SeatStatus": {
            "description": "Estado del asiento",
            "type": "string",
            "enum": [
                "AVAILABLE",
                "LOCKED",
                "SOLD"
            ],
            "x-enum-varnames": [
                "StatusAvailable",
                "StatusLocked",
                "StatusSold"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Formato: Bearer \u003ctoken\u003e\nOperaciones para gestionar eventos\nOperaciones para gestionar asientos\nOperaciones para gestionar órdenes de reserva\nOperaciones para gestionar el proceso de checkout",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}